Added Autonomouse Commands (NOT wired to DS)
Added the autonomouse commands along with a command that shoots the ball
and spins down the wheels.  This also works during teleop.  The auto and
spy auto groups are not hooked up to the smart dashboard, but that
should be as simple as a few edit boxes and retrieving it during
auto-init

+++ TinManXI/src/org/usfirst/frc/team1922/robot/commands/DriveForwardDistance.java
@@ -9,9 +9,9 @@ import edu.wpi.first.wpilibj.command.Command;
  */
 public class DriveForwardDistance extends Command {
 
-	protected final float mDistanceInches;
+	protected final double mDistanceInches;
 	
-    public DriveForwardDistance(float distanceInches) {
+    public DriveForwardDistance(double distanceInches) {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	mDistanceInches = distanceInches;
